+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerBaseConfiguration.kt
@@ -11,6 +11,7 @@ interface MatrixMessengerBaseConfiguration {
     var imprintUrl: String?
     var licenses: String?
     var pushUrl: String?
+    var multiProfile: Boolean
 
     fun copyTo(other: MatrixMessengerBaseConfiguration) {
         other.appName = appName
@@ -22,5 +23,6 @@ interface MatrixMessengerBaseConfiguration {
         other.imprintUrl = imprintUrl
         other.licenses = licenses
         other.pushUrl = pushUrl
+        other.multiProfile = multiProfile
     }
 }