+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/MatrixMessengerConfiguration.kt
@@ -53,6 +53,7 @@ data class MatrixMessengerConfiguration(
     override var licenses: String? = null,
 
     override var pushUrl: String? = null,
+    override var multiProfile: Boolean = true,
 
     /**
      * Inject and override modules.